+++ test.lisp
@@ -1,3 +1,6 @@
-( print ( + 1 2 ) )
-( print ( * 3 10 ) )
-( print ( + "hello " "world" ) )
+(print (-4))
+(print (- (-4) 2))
+(print (* 2 (-4)))
+
+"Use `neg` keyword to negate expressions"
+(print neg (+ 5 4))